Add SectionNote deletion spec

+++ app/controllers/api/v1/sections/section_notes_controller.rb
@@ -35,6 +35,9 @@ module Api
 
         def destroy
           section_note = section.section_note
+
+          raise Sequel::RecordNotFound if section_note.blank?
+
           section_note.destroy
 
           respond_with section_note
